4 osio paitsi 4.16 & 4.22

+++ blogilista/controllers/blogs.js
@@ -4,14 +4,14 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({}).populate('user', {username:1, name:1, id:1})
-      response.json(blogs.map(blog => blog = blog.toJSON()))
+  const blogs = await Blog.find({}).populate('user', { username:1, name:1, id:1 })
+  response.json(blogs.map(blog => blog = blog.toJSON()))
 })
 
 blogsRouter.post('/', async (request, response, next) => {
-  
+
   const body = request.body
- 
+
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!request.token || !decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
@@ -31,7 +31,7 @@ blogsRouter.post('/', async (request, response, next) => {
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
     response.json(savedBlog.toJSON())
-    } 
+  }
   catch(exception) {
     next(exception)
   }
@@ -54,13 +54,30 @@ blogsRouter.get('/:id', async (request, response, next) => {
 
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-  try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
   }
-  catch (exception) {
-    next(exception)
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if ( blog.user.toString() === user.id.toString() ) {
+    try {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    }
+    catch (exception) {
+      next(exception)
+    }
   }
+  else {
+    console.log('moi')
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
@@ -71,7 +88,7 @@ blogsRouter.put('/:id', async (request, response, next) => {
   }
 
   try {
-    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new:true})
+    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new:true })
     response.json(updatedBlog.toJSON())
   }
   catch(exception) {
